路桥2019/5/22 night

+++ src/barber/dao/HairDao.java
@@ -13,7 +13,7 @@ public class HairDao {
 
     //    查询长发的图片信息
     public static List<HairStyleBean> quaryLong() {
-        return quaryHairStyle("short");
+        return quaryHairStyle("long");
     }
 
     //    查询不同发型的图片信息
